refactor(desktop): decompose kanban page into models and subcomponents
